+++ src/main/java/com/yanny/ytech/configuration/block/TanningRackBlock.java
@@ -153,7 +153,7 @@ public class TanningRackBlock extends Block implements EntityBlock {
     }
 
     public static void registerRecipe(@NotNull RecipeOutput recipeConsumer, @NotNull DeferredItem<Item> item, MaterialType material) {
-        WorkspaceCraftingRecipe.Builder.recipe(item.get())
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.AXES.tag, item.get())
                 .define('W', Utils.getLogFromMaterial(material))
                 .define('T', YTechItemTags.GRASS_TWINES)
                 .bottomPattern("   ")